refactor(desktop): split app shell boundaries
