+++ apps/desktop/src/main/__tests__/renderer-shell-source-helpers.ts
@@ -8,6 +8,19 @@ const sourcePaths = [
   'main.tsx',
   'app.tsx',
   'app-shell.tsx',
+  'app-shell-plan-actions.ts',
+  'app-shell-skill-actions.ts',
+  'app-shell-project-actions.ts',
+  'app-shell-open-skill-action.ts',
+  'app-shell-import-actions.ts',
+  'app-shell-stop-action.ts',
+  'app-shell-copy.ts',
+  'app-shell-command-actions.ts',
+  'app-shell-effects.ts',
+  'app-shell-session-events.ts',
+  'app-shell-session-row-actions.ts',
+  'app-shell-session-settings-actions.ts',
+  'app-shell-visual-smoke.ts',
   'cached-theme-bootstrap.ts',
   'chat-model-selection.ts',
   'conversation-markdown.ts',
